+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 <!-- towncrier release notes start -->
 
+## 1.0.4
+
+- use default port naming strategy for sax.models
+
+## 1.0.3
+
+- fix models and add tests for models [#138](https://github.com/gdsfactory/cspdk/pull/138)
+
+## 1.0.2
+
+- update gdsfactory 9.9.4
+
 ## 0.16.0
 
 - update gdsfactory 9.4.0 [#110](https://github.com/gdsfactory/cspdk/pull/110)